Update and rename Weyl-Groups-2.js to Weyl-Groups.js
Major update; I reworked the main equation and vastly simplified the theory. I also reworked the logic of the theory to take advantage of the linear algebra underlying it. Much of the gating is now present, although no chapters or achievements, no localization, etc. Also the stability is not great and the balance is terrible (though it plays almost well up to publishing!), and although a player can probably figure out what's going on I would like to make it a bit of a softer introduction. I need to update the readme to reflect the changes