First pass at balancing
I can't make any guarantees for the whole theory, but I think it plays nicely for the first e20 and it seems like it plays reasonably at e100 (maybe a bit quickly). In particular, milestone, permanent, and letter upgrades are set how I want them, with possible minor adjustments in the future.
The next step is to make the theory more beginner-friendly. This means writing a more detailed and engaging description, as well as creating chapters and achievements, and a possible info button on the second page of the equation area.
For experts, I will introduce a "delete the last letter" button to fully preempt milestone swap strategies (at least the ones I can think of).
I may also adjust the logic (but not the cost) of letter buying to improve stability in the next update.
In-depth stability and UI improvements are coming, along with further balance improvements, but likely not in the next update.