Chapters and achievements
I implemented chapters and achievements. They may be buggy, and the exact points they trigger are subject to change. I also fixed a bug that made the letter auto-buyer useless, and added a "delete letter" button that works with some of the achievements and preempts a milestone swap strategy. Finally, I increased the publication multiplier to make the first e120 rho less of a slog.
The next update will be beta, and will have better balancing and an endgame feature. Following updates should all be minor